Issue 12644: Improve acmeCA FAT Lite bucket
