+++ dev/com.ibm.ws.security.acme/src/com/ibm/ws/security/acme/internal/util/AcmeConstants.java
@@ -93,5 +93,7 @@ public class AcmeConstants {
 
 	public static final Long SCHEDULER_MS = TimeUnit.HOURS.toMillis(24L);
 	public static final Long SCHEDULER_ERROR_MS = TimeUnit.HOURS.toMillis(1L);
+	
+	public static final String ACME_HISTORY_FILE = "acmeca-history.txt";
 
 }
